+++ client/order/src/main/java/com/live_commerce/order/kafkaOrder/payment/PaymentEventConsumer.java
@@ -1,5 +1,7 @@
 package com.live_commerce.order.kafkaOrder.payment;
 
+import com.live_commerce.order.application.service.OrderService;
+import com.live_commerce.order.kafkaOrder.service.PaymentSuccessServiceKafka;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
 
@@ -11,6 +13,8 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class PaymentEventConsumer {
 
+    private final PaymentSuccessServiceKafka paymentSuccessServiceKafka;
+
     //payment -> order
     private static final String COMPLETED_TOPIC = "payment-completed";
 
@@ -20,5 +24,6 @@ public class PaymentEventConsumer {
     )
     public void listenPaymentCompleted(PaymentCompletedEvent msg) {
         log.info("✅ 결제 성공 이벤트 수신(kafka): orderId={}, message={}", msg.orderId(), msg.message());
+        paymentSuccessServiceKafka.updatePaymentSuccessKafka(msg);
     }
 }
