+++ client/order/src/main/java/com/live_commerce/order/kafkaOrder/service/OrderServiceKafka.java
@@ -153,41 +153,6 @@ public class OrderServiceKafka {
         }
     }
 
-    //TODO KAFKA
-    //결제 처리 응답값 가져오기
-    @Transactional
-    public String updatePaymentSuccessKafka(PaymentCompletedEvent event){
-        Order order = orderRepository.findById(event.orderId())
-                .orElseThrow(() -> new OrderException(OrderExceptionCode.NOT_FOUND));
-        log.info("주문 들고오기 성공");
-        if(event.message() == null){
-            throw new OrderException("결제 상태가 COMPLETED가 아닌 상태입니다. 다시 결제해주세요 ", HttpStatus.FORBIDDEN);
-        }
-
-        //상태 변경 PAID로 변경
-        order.changeStatus(PAID);
-        log.info("READY 에서 PAID로 변경 성공!");
-
-        //TODO KAFKA 처리
-        //재고 감소 진행
-        //productClient.decreaseInventory(new InventoryDecreaseRequestDto(order.getProductId(), order.getProductQuantity()));
-        OrderCreatedEvent eventInventory = new OrderCreatedEvent(order.getId(), order.getProductId(), order.getProductQuantity());
-        inventoryEventProducer.sendOrderCreatedEvent(eventInventory);
-        log.info("재고 감소 성공!!");
-
-        //TODO KAFKA
-        // 7. 쿠폰 사용 처리
-//        if (order.getCouponId() != null) {
-//            couponClient.useCoupon(order.getCouponId());
-//        }
-        UUID userId = UUID.randomUUID();
-        if(order.getCouponId() != null){
-            CouponUsedMessage eventCoupon = new CouponUsedMessage(order.getCouponId(), userId);
-            couponUsedProducer.sendCouponUsedEvent(eventCoupon);
-        }
-        return "결제 성공 처리 완료";
-    }
-
     //kafka
     public void sendMessage(String topic , String key, String message) {
         for (int i = 0; i < 10; i++) {
