+++ client/order/src/main/java/com/live_commerce/order/kafkaOrder/service/PaymentStatusTransitionServiceKafka.java
@@ -15,6 +15,10 @@ import com.live_commerce.order.infrastructure.client.feign.ProductClient;
 import com.live_commerce.order.infrastructure.client.request.InventoryIncreaseRequestDto;
 import com.live_commerce.order.infrastructure.client.request.PaymentReadyRequestDto;
 import com.live_commerce.order.infrastructure.client.response.PaymentRefundResponseDto;
+import com.live_commerce.order.kafkaOrder.coupon.CouponUsedMessage;
+import com.live_commerce.order.kafkaOrder.coupon.CouponUsedProducer;
+import com.live_commerce.order.kafkaOrder.payment.PaymentCompletedEvent;
+import com.live_commerce.order.kafkaOrder.product.OrderCreatedEvent;
 import com.live_commerce.order.presentation.common.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -25,16 +29,22 @@ import org.springframework.transaction.annotation.Transactional;
 import java.math.BigDecimal;
 import java.util.UUID;
 
+import static com.live_commerce.order.domain.model.OrderStatus.PAID;
+
 @Slf4j
 @Service
 @RequiredArgsConstructor
 //결제 상태 변경
 public class PaymentStatusTransitionServiceKafka {
+
     private final ProductClient productClient;
     private final PaymentClient paymentClient;
     private final CouponClient couponClient;
     private final OrderRepository orderRepository;
     boolean changeChangeStatus = false;
+
+    //KAFKA
+    private CouponUsedProducer couponUsedProducer;
 
     //pending -> ready -> paid
