+++ server/config/src/main/resources/config-repo/order-dev.yml
@@ -27,7 +27,8 @@ spring:
       value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
       properties:
         spring.json.type.mapping: >
-          payment-completed:com.live_commerce.order.kafkaOrder.payment.PaymentCompletedEvent
+          payment-completed:com.live_commerce.order.kafkaOrder.payment.PaymentCompletedEvent,
+          coupon-used:com.live_commerce.order.kafkaOrder.coupon.CouponUsedMessage
     consumer:
       group-id: ${spring.application.name}
       auto-offset-reset: earliest
@@ -36,7 +37,8 @@ spring:
       properties:
         spring.json.trusted.packages: '*'
         spring.json.type.mapping: >
-          payment-completed:com.live_commerce.order.kafkaOrder.payment.PaymentCompletedEvent
+          payment-completed:com.live_commerce.order.kafkaOrder.payment.PaymentCompletedEvent,
+          coupon-used:com.live_commerce.order.kafkaOrder.coupon.CouponUsedMessage
 
     listener:
       ack-mode: MANUAL_IMMEDIATE
